compactLogs removes all logs in the snapshot
Previously we kept the last log in the snapshot for the purposes
of replication to avoid an edge case. However, this introduces a
potential bug as the last snapshot on a follower may be in conflict
and must be removed. We now compact that log, but have to make an
appropriate check in the replication. takeSnapshot must also set
the latest information before compacting logs. There is a potential
race between the snapshot/replication routines. This can cause an
AppendEntries command to fail, but it would have likely failed anyways
since it is trying to read a log that is being compacted. The basic
retry of this should resolve it.