Removing optimization around appendEntries RPC
The optimization avoids an appendEntries call when the follower is
up-to-date. However, if a follower restarts, its log may be complete
but the commit index will be 0 until the next appendEntries call. This
can potentially be a very long time if there is no activity, which
results in stale results on the followers.