Allow users to pass in their own logger.
For making raft log to github.com/golang/glog, it is necessary to
specify a custom logger, because glog expects you to set the Lshortfile
flag.

Aside from this specific example, I think in general it’s better to give
users more control over the logger instead of creating it for the user —
after all, it’s still just a single line of code.