Close the transport on shutdown
When a peer shuts down, close the transport. This needs to be done in
the future (i.e. when all the routines have ended), else the remove
peer does not propagate correctly - this was an issue with pull
request #72.

Unfortunately we cannot simply add a Close() to the Transport
interface as we would break back compatibility. Therefore
add a type assertion to ensure that the interface supports
Close before closing it.

Add a Close() method to InmemTransport so this actually gets
tested. InmemTransport has nothing much to close (though it
does call DisconnectAll), but other transports may do.

Signed-off-by: Alex Bligh <alex@alex.org.uk>