Change calculation of heartbeat timer reset
According to the raft spec, we should time out between 1 and 2 intervals
(here called r.conf.HeartbeatTimeout) after the last contact. Before this
change, the heartbeat timeout was set to between 1 and 2 intervals after
Now(). Now() may be substantially after the last contact as it itself
is driven by a heartbeat timeout.

There is a possibility that this value (1 and 2 intervals since
last contact) is before Now(). So what we do is take a randomized interval
from the later of the last interval and Now() - r.conf.HeartbeatTimeout.

Signed-off-by: Alex Bligh <alex@alex.org.uk>