Stop truncating non-conflicting entries in AppendEntries
The AppendEntries request handler used to temporarily truncate entries
from the store that had indexes overlapping with the request, then it'd
go back and add all the entries in the request. This opened a window
of vulnerability during which if the server crashed, it could have
deleted some log entries that it shouldn't have.

For this to happen, the follower would have to receive an AppendEntries
request with entries that it already had. In principle, that could be
due to a duplicate request (depends on transport behavior) or to a
leader with a nextIndex set too low (probably isn't supposed to happen
with the back-up-one-entry algorithm).

The new code truncates only conflicting entries from the log store, then
appends only the new entries.

This also fixes a related issue where the last log index and term are
set to the last entry in the request, even if the last entry in the
request is not the last entry in the log (suppose the request contained
an earlier subsequence of the existing log).