Track peer IDs.
* encode/decode ServerID in transports
* track the leader ServerID
* use the leader ServerID as a second identifier

Votes will be rejected when `candidate(id,addr)` doesn't match up with
`leader(id,addr)`. The reasoning here is that a change of the
ServerID or the ServerAddress indicates a change in the peer process
(This is especially important in more dynamic environments like
kubernetes where address can be reused by different nodes).