Fix deadlock between Shutdown() and takeSnapshot()
This change makes sure that a takeSnapshot() call which is run about
at the same time as Shutdown() does not block indefinitely on a
ConfigurationFuture request.

The situation can happen if takeSnapshot() manages to push the request
to configurationCh (which is buffered) but the
follower/candidate/leader loop exists before processing it, in turn
making the shutdownFuture.Error() returned by Shutdown() block
indefinitely (since it indirectly waits for takeSnapshot to finish).

With this change the sample program attached to #268 doesn't fail
anymore.