Make TestNetworkTransport_CloseStreams less flaky
This change improves the robustness of the
TestNetworkTransport_CloseStreams test, which might occasionally fail
on Travis (as in [1]) if Transport 1 consumes RPCs fast enough to not
stress Transport 2's pool as expected.

The test is still flaky in principle, but I belive it should be good
enough to make CI green most (all?) the times.

[1] https://travis-ci.org/hashicorp/raft/jobs/340532532