[FIXED] LogCache should not cache on StoreLogs error
LogCache was caching logs and then invoking store.StoreLogs().
However, if StoreLogs() fails to store, LogCache.Get() could still
return logs that had not been persisted into storage, which could
lead to a node failing to restart with a "log not found" panic.

This PR ensures that LogCache only cache logs if StoreLogs was
successful.

Resolves #429

Signed-off-by: Ivan Kozlovic <ivan@synadia.com>