testing: improving logging output for tests
A flaky test failed in CI and I noticed the log output was difficult to
follow. I believe the reason is that testLoggerAdapter was double
printing the logs.

In CI tests are run using `gotestsum` which runs `go test -json`. The
`-json` flag implies `-v`, so in CI tests are always run in verbose
mode. `testLoggerAdapter` was checking `testing.Verbose` and priting
logs twice, once to stdout, and once to `t.Log`.

Instead of double printing, we can handle this logic in `newTestLogger`
helpers. If verbose mode is enabled we can use a regular logger to get
real-time output of the logs. In "quiet" mode we continue to use `t.Log`
as the output for logs so that logs are only printing on failure.

Also removed a couple duplicate helper functions. Setting `Level` to
the default is a no-op, so many of these functions were actually doing
the same thing.

It is important to note that in general sending logs to t.Log can be a
problem because if a goroutine outlives the main test function it will
cause a panic. This does not appear to be a problem right now, so I have
not changed it.