Add a gauge to hold the last restore time that
This is updated on every restore including startup from disk as well as RPC (leader/user initiated restore).

It's a gauge to provide a constant reading of the last restore time which may have been days or weeks ago in a healthy server but is useful to be able to graph along side the trailing logs age to check if there is danger of restores taking longer than trailing logs can cover.

It differes from fsm.restore both because it is omitted at startup and because it is a gauge rather than a single point in time sample.