Fix flaly TestRaft_LeadershipTransferLeaderRejectsClientRequests
This test is currently the test which flakes the most often.

The test would fail because of the select from channel. This select race
with the `leaderLoop` goroutine. If the read-from-channel in the test won
then the `leaderLoop` goroutine would never receive the future, and the
future would never respond, leaving the test to timeout eventually.

The fix was to remove the select and read-from-channel from the test. This fix
caused the test to fail, because of the following call to `LeadershipTransferToServer`
would not being able to send to the channel immediately, resulting in an
ErrEnqueueTimeout instead of ErrLeadershipTransferInProgress.

To fix the secondary problem the test was changed to call
`LeadershipTransferToServer` after the channel is drained (by waiting on
all the futures).

Also added a `t.Logf` to the test, since that helped debug the problem I
thought it would be good to keep it around in case we encounter other
problems with this test.