Fix new 'go vet' errors reported by Go1.16
Calling Fatalf in a goroutine is a problem because Fatalf calls FailNow, which calls runtime.Goexit.
Goexit stops the goroutine, but nothing else. When called from a goroutine that is not the main test
Goroutine this does not have the desired affect. It does not stop the test from running.

Use Errorf and return for now. In the future we may want to replace these with a channel that sends
errors back to the main goroutine.