Fix peer access race with raft heartbeat.
This race recently came up when the Juju project moved to the upstream
raft project. The problem was manifested during testing
(unit/integration test) setup. The race happens during a Raft
replicate, a goroutine is spawned to manage the heartbeats against a
peer. The problem occurs when a peer can be changed during a
startStopReplication call. If the server.ID already exists, yet the
server address isn't the same, then the underlying peer is updated and
a data race occurs.

The fix is to wrap all peer access with a RWMutex. As the pattern in the
library uses a lot more reads than writes, it makes much more sense to
use this type of mutex.