Pre-Vote Campaign: Initial thoughts on pre-voting
This is a very rough draft on adding pre-voting (see 9.6 of the raft
thesis) to raft. The commit is just familiarising the internal code base
before ensuring the correct implementation.

In the Pre-Vote algorithm, a candidate only increments its term if it
first learns from a majority of the cluster that they would be willing
to grant the candidate their votes. Thus saving any potential disruption
when a server attempts to rejoin.