Existing tests with PreVoting
To ensure we haven't regressed any existing tests, we run the full
test suite with PreVoting campaign off and on. Although not ideal,
any new and old tests are run with both PreVoting on and off. It
would be better to have some sort of test suite for this kind of
thing, yet we can replicate that same behaviour here.