Make sure leadership transfer is set back to false in all cases.
Release 1.3.4 introduced a bug that wound only call setLeadershipTransferInProgress back to false
when leadership transfer to another node succeeded.  In the case where it did not succeed,
leadershipTransferInProgress was left set to true, breaking any callers that relied on it being set back
to false after a failed leadership transfer.