Introduce MonotonicLogStore interface
This commit introduces an interface which acts as a handler for a leaky
abstraction in the structure of underlying log stores. In order to
properly handle post-snapshot-restore cleanup for log stores
generically, we need some awareness of whether the underlying store
permits gaps.

Boltdb allows for gaps in log store indices, but to handle them it
requires a freelist, which is written on every commit. This is costly,
particularly when the freelist is large. By completely resetting the
LogStore after snapshot, we grow the size of the freelist, which would
result in performance degradation.

The MonotonicLogStore interface is implemented by LogStores with
guarantees of sequential/monotonic indices, like raft-wal, but reverts
to the old behavior for boltdb.

This also requires special handling within LogStore wrappers (like
LogCache), to ensure that the type assertion is passed to the underlying
store.