Upgrade go-msgpack to 2.1.1
We add a parameter that allows the behavior of the time encoding to be
changed by `NetworkTransportConfig.MsgpackUseNewTimeFormat`.
The default value is `false`, which keeps the current behavior of time
encoding is compatible with with `go-msgpack` v0.5.5 (the version
in `go.mod` prior to this change).

However, users who upgrade to this version of Raft will no longer
have their own version of pre-v2 `go-msgpack` honored, which could
cause unexpected binary compatibility problems.

This will be a problem for HashiCorp products who might blindly
upgrade:
* [Vault](https://github.com/hashicorp/vault/blob/f2f532ec802d356346c2302e9c38734f2a594f3f/go.mod#L98)
  (`go-msgpack` v1.1.5)
* [Nomad](https://github.com/hashicorp/nomad/blob/cb2363f2fbbb771af9c8a62ed09a60250df647a2/go.mod#L64)
  (`go-msgpack` v1.1.5).

I don't believe this will be a problem for [Consul](https://github.com/hashicorp/consul/blob/8eb074e7c17a0a2077436637c06f521b036407f5/go.mod#L199)
(`go-msgpack` v0.5.5).

For the sake of non-HashiCorp products using this library, I'd recommend
merging this and making a release with a full `v2` so that people are
not surprised.

I'll follow-up this PR and release with PRs to upgrade Vault, Consul,
and Nomad.

I tested this branch with Vault using
`NetworkTransportConfig.MsgpackUseNewTimeFormat` set to `true` and
verified it was able to replicate to Vault 1.15.0 (which uses
`go-msgpack` v1.1.5). (I also tested that the same setting of `false`
would fail to replicate to Vault 1.15.0 with a time decoding error,
as expected).

I also added a quick benchmark in `bench_test.go`, and did not notice
any significant performance differences before and after this change
(which is not surprising, since I don't assume msgpack is a significant
cost for snapshots).