fix(trans): Reset all encoders when using pooled conns.
We have observed at customer sites various hard-to-replicate cluster
instabilities where followers will fail to resync with the rest of the
cluster after a restart.  In particular, we get a couple of different
types of errors from Raft:

failed to decode incoming command: error=msgpack decode error [pos 12207]: only encoded map or array can be decoded into a struct

This one generally happens when the follower is far enough behind that
the leader decides to send a snapshot instead.

The second is one that should not be possible:

failed to decode incoming command: error=unknown rpc type 125

We have seen various different RPC types with different numbers
that do not correspond to any of the 5 known Raft RPC types that this
package implements.

On the leader side, we want to nail shut the possibility that any sort of
conn buffer reuse might be the cause of these errors.  This patch forces
all pooled conns to reset their buffers and encoders whenever a conn
is added to or removed from its transport, and arranges for things
to panic if a pooled conn is used when it should not be.