debug(buffers): Enhance debugging of RPC recieve failures.
We have observed at customer sites various hard-to-replicate cluster
instabilities where followers will fail to resync with the rest of the
cluster after a restart.  In particular, we get a couple of different
types of errors from Raft:

failed to decode incoming command: error=msgpack decode error [pos 12207]: only encoded map or array can be decoded into a struct

This one generally happens when the follower is far enough behind that
the leader decides to send a snapshot instead.

The second is one that should not be possible:

failed to decode incoming command: error=unknown rpc type 125

We have seen various different RPC types with different numbers
that do not correspond to any of the 5 known Raft RPC types that this
package implements.

On the follower side, assist with debugging these sorts of errors by
printing a short snippet (up to 100 bytes) of the remaining data in the
connection's read buffer whenever any RPC is unable to be completed
and there is any remaining unread buffered data.