fix(snapshots): Fix snapshot handling bugs.
We have observed at customer sites various hard-to-replicate cluster
instabilities where followers will fail to resync with the rest of the
cluster after a restart.  In particular, we get a couple of different
types of errors from Raft:

failed to decode incoming command: error=msgpack decode error [pos 12207]: only encoded map or array can be decoded into a struct

This one generally happens when the follower is far enough behind that
the leader decides to send a snapshot instead.

The second is one that should not be possible:

failed to decode incoming command: error=unknown rpc type 125

where the RPC number seems to change almost at random.

After many rounds of debugging at a remote customer site, it appears
that the root cause of the issue is that the Raft library will send
extra data if the file size of the snapshot on disk is larger than the
snapshot size indicated in the snapshot metadata.  In my particular
usecase that can happen because the SnapshotStore we use prefers to
recycle older snapshot files rather than deleting them and creating a
new file.  We do this to make the system more resistant to running out
of disk space.

On the leader side, alleviate this issue by clamping the maximum
amount of snapshot data we will send to the size indicated by the
snapshot metadata.

On the follower side, rework the RPC event loop to exit after handling
an installSnapshot RPC instead of attempting to process the rest of
the bytes remaining in the conn.  This mirrors what the leader has
done for snapshot handling with pooled conns since forever.

On the leader side, arrange for pooled connections to discard any
outstanding data when a connection is returned to the pool, and
arrange for the system to panic if someone ever attempts to use a conn
after returning it to the pool.