fix(snapshot): Fix a few errors in snapshot handling.
On the server side, it was possible for the InstallSnapshot RPC to send more
data than the client was expecting to handle.  This can happen when the
io.Reader containing the snapshot can return more data than the snapshot
metadata says it contains. Fix this by using a LimitedReader to clamp the
amount of snapshot data that the Reader will send to that which the metadata
says it should expect.

On the client side, it was possible for the server side error to result in
handleCommand returning a spurious error that could lead to the follower
getting out of sync with the leader in a variety of interesting ways.